regression 4216: libsks can guesses key type
libsks can guess the key size when key generation mechanism is
CKM_RSA_PKCS_KEY_PAIR_GEN. This feature is implemented in libsks
to match generic tools as pkcs11-tool.

This change modifies the test to have key type defined in the
public key attributes but not in the private key attribute. Libsks
shall guess key type for the later.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>